ansible fix

+++ ansible.sh
@@ -10,7 +10,7 @@ ansible-galaxy install elastic.elasticsearch,5.5.1
 
 # test to see if we've linked this in development
 # install by default in production
-if [ ! -e roles/csirtgadgets.cif ]; then
+if [ ! -e roles/csirtgadgets.cif ] && [ ! -e ../roles/csirtgadgets.cif ]; then
   ansible-galaxy install csirtgadgets.cif
 fi
